+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 - Pass an `$initiator` into `ResponseDecoderInterface::supports` method
 - Fix stopwatch problem when calling a child query/command
+- Fix profiling used decoders when calling a child query/command
 
 ## 1.0.0 - 2021-05-13
 - Initial implementation